fix: some refactoring
